requirements.txt
beautifulsoup4
pandas
python-slugify
requests
